@@ -32,25 +32,30 @@
     "*.{mjs,js,ts,svelte,css,md,json}": "prettier --cache --write"
   },
   "devDependencies": {
-    "@types/node": "^18.13.0",
-    "auto": "^10.40.0",
-    "commitlint": "^17.4.2",
+    "@types/node": "^22.13.9",
+    "auto": "^11.3.0",
+    "commitlint": "^19.8.0",
     "commitlint-config-unperfekt": "workspace:*",
-    "eslint": "^8.33.0",
-    "husky": "^8.0.3",
-    "lerna": "^6.4.1",
-    "lint-staged": "^13.1.1",
-    "prettier": "^2.8.4",
+    "eslint": "^9.21.0",
+    "husky": "^9.1.7",
+    "lerna": "^8.2.1",
+    "lint-staged": "^15.4.3",
+    "prettier": "catalog:",
     "prettier-config-unperfekt": "workspace:*",
-    "sort-package-json": "^2.4.1",
-    "turbo": "^1.7.4"
+    "sort-package-json": "^3.0.0",
+    "turbo": "^2.4.4"
   },
-  "packageManager": "pnpm@7.27.0",
+  "packageManager": "pnpm@10.6.1",
   "engines": {
     "node": ">=16.13.0"
   },
   "publishConfig": {
     "access": "public",
     "registry": "https://registry.npmjs.org/"
+  },
+  "pnpm": {
+    "onlyBuiltDependencies": [
+      "esbuild"
+    ]
   }
 }
