@@ -24,23 +24,26 @@
     "README.md"
   ],
   "dependencies": {
-    "@typescript-eslint/eslint-plugin": "^5.46.1",
-    "@typescript-eslint/parser": "^5.46.1",
-    "eslint": "^8.29.0",
-    "eslint-config-prettier": "^8.5.0",
-    "eslint-plugin-import": "^2.26.0",
-    "eslint-plugin-jsdoc": "^39.6.4",
-    "eslint-plugin-prettier": "^4.2.1",
-    "eslint-plugin-svelte3": "^4.0.0",
-    "prettier": "^2.8.1",
-    "svelte": "^3.55.0",
-    "typescript": "^4.9.4"
+    "@eslint/compat": "^1.2.5",
+    "@eslint/js": "^9.18.0",
+    "@typescript-eslint/eslint-plugin": "^8.26.0",
+    "@typescript-eslint/parser": "^8.26.0",
+    "eslint": "^9.21.0",
+    "eslint-config-prettier": "^10.1.1",
+    "eslint-plugin-import": "^2.31.0",
+    "eslint-plugin-jsdoc": "^50.6.3",
+    "eslint-plugin-svelte": "^3.0.0",
+    "globals": "^16.0.0",
+    "prettier": "catalog:",
+    "svelte": "catalog:",
+    "typescript": "catalog:",
+    "typescript-eslint": "^8.20.0"
   },
   "peerDependencies": {
     "eslint": "^8",
-    "prettier": "^2",
-    "svelte": "^3",
-    "typescript": "^4"
+    "prettier": "catalog:",
+    "svelte": "catalog:",
+    "typescript": "catalog:"
   },
   "peerDependenciesMeta": {
     "eslint": {
@@ -56,7 +59,6 @@
       "optional": true
     }
   },
-  "packageManager": "pnpm@7.27.0",
   "publishConfig": {
     "access": "public",
     "registry": "https://registry.npmjs.org/"
