@@ -1,4 +1,4 @@
-module.exports = {
+export default {
   // extends: ["gitmoji"],
   // extends: ["commitlint-config-gitmoji"],
   extends: ["@commitlint/config-conventional"],
