@@ -1,5 +1,7 @@
-module.exports = {
-  extends: ["plugin:jsdoc/recommended"],
+import jsdoc from 'eslint-plugin-jsdoc';
+
+export default {
+  ...jsdoc.configs['flat/recommended'],
   rules: {
     "jsdoc/require-description-complete-sentence": "error",
     "jsdoc/require-param": "off",
