@@ -1,8 +1,5 @@
-module.exports = {
+export default {
   plugins: {
-    "postcss-import": {},
-    "tailwindcss/nesting": "postcss-nesting",
-    "tailwindcss": {},
-    "autoprefixer": {},
+    "@tailwindcss/postcss": {}
   },
 }
