@@ -1,7 +1,4 @@
 /** @type {import('@types/tailwindcss/tailwind-config').TailwindConfig} */
-module.exports = {
-  corePlugins: {
-    float: false,
-  },
-  plugins: [require("@tailwindcss/typography")],
+export default {
+  plugins: ["@tailwindcss/typography"],
 }
