@@ -1,4 +1,6 @@
-<script lang="ts" strictEvents>
+<svelte:options runes={true} />
+
+<script lang="ts">
   import { setContext } from "svelte"
 
   import { createSelectionManager } from "../../../state/collection/selection.js"
@@ -10,14 +12,18 @@
   import type { AccordionContext } from "./context.js"
   import type { Key } from "../../../state/collection/selection.js"
 
-  /** Wheter to allow multiple items to be open at the same time or not. */
-  export let multiple: boolean = false
+  interface AccordionProps {
+    /** Wheter to allow multiple items to be open at the same time or not. */
+    multiple?: boolean
+    /** The items to display in the accordion. */
+    items?: Item[]
+    /** The keys of the items that are currently selected. */
+    selectedKeys?: Set<Key>
+    /** The children to render if no items are provided. */
+    children: () => any
+  }
 
-  /** Item objects in the collection. */
-  export let items: Item[] = []
-
-  /** The currently selected keys in the collection. */
-  export let selectedKeys: Set<Key> = new Set<Key>()
+  let { multiple = false, items = [], selectedKeys = new Set(), children }: AccordionProps = $props()
 
   const selectionStore = createSelectionManager({
     multiple,
@@ -27,12 +33,12 @@
   setContext<AccordionContext>(contextKey, selectionStore)
 </script>
 
-<div on:click class="Accordion">
+<div on:click on:keydown class="Accordion">
   {#each items as item (item.id)}
     <AccordionItem key={item.id} title={item.title}>
       {item.description}
     </AccordionItem>
   {:else}
-    <slot />
+    {@render children()}
   {/each}
 </div>
