@@ -1,17 +1,20 @@
-<script lang="ts" strictEvents>
+<script lang="ts" >
   /**
    * @name ButtonGroup
    *
    * ButtonGroup is a container for buttons.
    */
   import { default as cn } from "classnames"
+  interface ButtonGroupProps {
+    /** Space-separated list of the case-sensitive classes of the element. */
+    class?: string
+    /** Whether the buttons are disabled. */
+    disabled?: boolean
+    /** The children to render inside the button group. */
+    children?: () => any
+  }
 
-  /** Space-separated list of the case-sensitive classes of the element. */
-  let _class: string | undefined = undefined
-  export { _class as class }
-
-  /** Whether the buttons are disabled. */
-  export let disabled: boolean | undefined = undefined
+  let { class: _class, disabled, children, ...restProps }: ButtonGroupProps = $props()
 </script>
 
 <div
@@ -22,8 +25,9 @@
     },
     _class,
   )}
-  disabled={disabled}
-  {...$$restProps}
+  {...restProps}
 >
-  <slot />
+  {#if children}
+    {@render children()}
+  {/if}
 </div>
